dev-libs/libmpack: fix build on Darwin and keyword

+++ dev-libs/libmpack/libmpack-1.0.5.ebuild
@@ -11,7 +11,7 @@ SRC_URI="https://github.com/${PN}/${PN}/archive/${PV}.tar.gz -> ${P}.tar.gz"
 
 LICENSE="MIT"
 SLOT="0"
-KEYWORDS="amd64 ~arm ~arm64 x86"
+KEYWORDS="amd64 ~arm ~arm64 x86 ~x64-macos"
 
 DEPEND=""
 RDEPEND="${DEPEND}"
@@ -30,6 +30,7 @@ src_prepare() {
 
 src_compile() {
 	local myemakeargs=(
+		"PREFIX=/usr"
 		"CC=$(tc-getCC)"
 		"config=release"
 		"LIBDIR=/usr/$(get_libdir)"
@@ -51,6 +52,11 @@ src_install() {
 	)
 
 	emake "${myemakeargs[@]}" install
+
+	if [[ ${CHOST} == *-darwin* ]] ; then
+        local file="libmpack.0.0.0.dylib"
+        install_name_tool -id "${EPREFIX}/usr/$(get_libdir)/${file}" "${ED}/usr/$(get_libdir)/${file}" || die "Failed to adjust install_name"
+    fi
 
 	find "${ED}" -name '*.la' -delete || die
 }
